ch(travis) refactor codes to fix travis error Fix travis build Refactor the getone account endpoint File folder structure change with addition of server
[Finishes #165593726]
