--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "workbench.colorCustomizations": {
-        "activityBar.background": "#5B0930",
-        "titleBar.activeBackground": "#800C43",
-        "titleBar.activeForeground": "#FFFCFD"
-    }
-}
