+++ src/controllers/accountController.js
@@ -1,3 +1,4 @@
+
 import moment from 'moment';
 
 import db from '../models/index';
@@ -29,10 +30,6 @@ class AccountControl {
   }
 
   static async getOne(req, res, next) {
-    if (req.decoded.type === 'client') {
-      res.status(401);
-      return next(new Error('Only staff and admins can view a specific account'));
-    }
     const {
       accountNumber,
     } = req.params;
@@ -59,26 +56,14 @@ class AccountControl {
       accountNumber,
     } = req.params;
     try {
-      const {
-        rows,
-      } = await db.query('SELECT * FROM accounts WHERE accountNumber=$1', [accountNumber]);
-      if (!rows[0]) {
-        res.status(404);
-        return next();
-      }
-      try {
-        const { rows } = await db.query('SELECT * FROM transactions WHERE accountNumber=$1', [accountNumber]);
-        res.json({
-          status: 200,
-          data: rows,
-        });
-      }
-      catch (e) {
-        next(e);
-      }
-    } catch (e) {
-      res.status(404);
-      next(new Error('Invalid Account Number'));
+      const { rows } = await db.query('SELECT * FROM transactions WHERE accountNumber=$1', [accountNumber]);
+      res.json({
+        status: 200,
+        data: rows,
+      });
+    }
+    catch (e) {
+      next(e);
     }
   }
